+++ driver_zodiac.c
@@ -239,7 +239,7 @@ static gps_mask_t handle1002(struct gps_device_t *session)
 	if (status & 1)
 	    session->gpsdata.satellites_used++;
 
-	session->gpsdata.skyview[i].PRN = prn;
+	session->gpsdata.skyview[i].PRN = (short)prn;
 	session->gpsdata.skyview[i].ss = (float)getzword(17 + (3 * i));
 	session->gpsdata.skyview[i].used = (bool)(status & 1);
     }
@@ -278,13 +278,13 @@ static gps_mask_t handle1003(struct gps_device_t *session)
 
     for (i = 0; i < ZODIAC_CHANNELS; i++) {
 	if (i < session->gpsdata.satellites_visible) {
-	    session->gpsdata.skyview[i].PRN = (int)getzword(15 + (3 * i));
+	    session->gpsdata.skyview[i].PRN = (short)getzword(15 + (3 * i));
 	    session->gpsdata.skyview[i].azimuth =
-		(int)(((short)getzword(16 + (3 * i))) * RAD_2_DEG * 1e-4);
+		(short)(((short)getzword(16 + (3 * i))) * RAD_2_DEG * 1e-4);
 	    if (session->gpsdata.skyview[i].azimuth < 0)
 		session->gpsdata.skyview[i].azimuth += 360;
 	    session->gpsdata.skyview[i].elevation =
-		(int)(((short)getzword(17 + (3 * i))) * RAD_2_DEG * 1e-4);
+		(short)(((short)getzword(17 + (3 * i))) * RAD_2_DEG * 1e-4);
 	} else {
 	    session->gpsdata.skyview[i].PRN = 0;
 	    session->gpsdata.skyview[i].azimuth = 0;
